requirements.txt
tensorflow==2.16.1
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.1
matplotlib==3.9.0
seaborn==0.13.2
Pillow==10.4.0
fastapi==0.112.0
uvicorn==0.30.5
python-multipart==0.0.9
streamlit==1.37.1
plotly==5.23.0
requests==2.32.3
locust==2.31.5
